% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network-core.R
\name{prn_counterdirection}
\alias{prn_counterdirection}
\title{Calculate counterdirectional regulatory input}
\usage{
prn_counterdirection(object)
}
\arguments{
\item{object}{A `prn` object.}
}
\value{
Target-by-comparison decomposition table.
}
\description{
Calculate counterdirectional regulatory input
}
