% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network-core.R
\name{prn_summary}
\alias{prn_summary}
\title{Summarize a PRN object}
\usage{
prn_summary(object)
}
\arguments{
\item{object}{A `prn` object.}
}
\value{
Comparison-level diagnostics.
}
\description{
Summarize a PRN object
}
